+++ loudmon/Source/loudmon/plot.cpp
@@ -15,10 +15,10 @@ void PlotComponent::paint_coordinates(Graphics &g) {
     auto [original_x, original_y] = map_point_reverse(x, y);
     // x axis
     g.fillEllipse(x-dot_radius/2, app_y_max-dot_radius/2, dot_radius, dot_radius);
-    g.drawSingleLineText(compact_value_text(original_x), x, app_y_max+pad/2+font_height/2, Justification::horizontallyCentred);
+    g.drawSingleLineText(compact_value_text(original_x), static_cast<int>(x), static_cast<int>(app_y_max+pad/2+font_height/2), Justification::horizontallyCentred);
 
     // y axis
     g.fillEllipse(app_x_min-dot_radius/2, y-dot_radius/2, dot_radius, dot_radius);
-    g.drawSingleLineText(compact_value_text(original_y), app_x_min-pad/2, y+font_height/2, Justification::right);
+    g.drawSingleLineText(compact_value_text(original_y), static_cast<int>(app_x_min-pad/2), static_cast<int>(y+font_height/2), Justification::right);
   }
 }
